Deprecate: initial LightningLoggerBase to Logger, tests/loggers pass
TODOs:
- add deprecation warning test
-  update docs
