+++ pytorch_lightning/loggers/__init__.py
@@ -13,11 +13,12 @@
 # limitations under the License.
 from os import environ
 
-from pytorch_lightning.loggers.base import LightningLoggerBase, LoggerCollection
+from pytorch_lightning.loggers.base import LightningLoggerBase
 from pytorch_lightning.loggers.csv_logs import CSVLogger
+from pytorch_lightning.loggers.logger import Logger, LoggerCollection
 from pytorch_lightning.loggers.tensorboard import TensorBoardLogger
 
-__all__ = ["LightningLoggerBase", "LoggerCollection", "TensorBoardLogger", "CSVLogger"]
+__all__ = ["Logger", "LoggerCollection", "TensorBoardLogger", "CSVLogger", "LightningLoggerBase"]
 
 from pytorch_lightning.loggers.comet import _COMET_AVAILABLE, CometLogger  # noqa: F401
 from pytorch_lightning.loggers.mlflow import _MLFLOW_AVAILABLE, MLFlowLogger  # noqa: F401
