+++ pytorch_lightning/loggers/csv_logs.py
@@ -27,7 +27,7 @@ from typing import Any, Dict, Optional, Union
 import torch
 
 from pytorch_lightning.core.saving import save_hparams_to_yaml
-from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_experiment
+from pytorch_lightning.loggers.logger import Logger, rank_zero_experiment
 from pytorch_lightning.utilities.logger import _add_prefix, _convert_params
 from pytorch_lightning.utilities.rank_zero import rank_zero_only, rank_zero_warn
 
@@ -100,7 +100,7 @@ class ExperimentWriter:
             writer.writerows(self.metrics)
 
 
-class CSVLogger(LightningLoggerBase):
+class CSVLogger(Logger):
     r"""
     Log to local file system in yaml and CSV format.
