+++ tests/deprecated_api/test_remove_1-7.py
@@ -242,9 +242,7 @@ def test_v1_7_0_deprecate_add_get_queue(tmpdir):
 
 def test_v1_7_0_lightning_logger_base_close(tmpdir):
     logger = CustomLogger()
-    with pytest.deprecated_call(
-        match="`LightningLoggerBase.close` method is deprecated in v1.5 and will be removed in v1.7."
-    ):
+    with pytest.deprecated_call(match="`Logger.close` method is deprecated in v1.5 and will be removed in v1.7."):
         logger.close()
     with pytest.deprecated_call(
         match="`LoggerCollection.close` method is deprecated in v1.5 and will be removed in v1.7."
