+++ tests/deprecated_api/test_remove_1-8.py
@@ -596,9 +596,9 @@ def test_v1_8_0_deprecated_agg_and_log_metrics_override(tmpdir):
 
     # Test single loggers
     with pytest.deprecated_call(
-        match="`LightningLoggerBase.agg_and_log_metrics` is deprecated in v1.6 and will be removed"
-        " in v1.8. `Trainer` will directly call `LightningLoggerBase.log_metrics` so custom"
-        " loggers should not implement `LightningLoggerBase.agg_and_log_metrics`."
+        match="`Logger.agg_and_log_metrics` is deprecated in v1.6 and will be removed"
+        " in v1.8. `Trainer` will directly call `Logger.log_metrics` so custom"
+        " loggers should not implement `Logger.agg_and_log_metrics`."
     ):
         Trainer(logger=logger)
     # Should have no deprecation warning
@@ -606,9 +606,9 @@ def test_v1_8_0_deprecated_agg_and_log_metrics_override(tmpdir):
 
     # Test multiple loggers
     with pytest.deprecated_call(
-        match="`LightningLoggerBase.agg_and_log_metrics` is deprecated in v1.6 and will be removed"
-        " in v1.8. `Trainer` will directly call `LightningLoggerBase.log_metrics` so custom"
-        " loggers should not implement `LightningLoggerBase.agg_and_log_metrics`."
+        match="`Logger.agg_and_log_metrics` is deprecated in v1.6 and will be removed"
+        " in v1.8. `Trainer` will directly call `Logger.log_metrics` so custom"
+        " loggers should not implement `Logger.agg_and_log_metrics`."
     ):
         Trainer(logger=[logger, logger3])
     # Should have no deprecation warning
