+++ tests/utilities/test_cli.py
@@ -33,7 +33,7 @@ from torch.optim.lr_scheduler import ReduceLROnPlateau, StepLR
 
 from pytorch_lightning import Callback, LightningDataModule, LightningModule, Trainer
 from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
-from pytorch_lightning.loggers import LightningLoggerBase, TensorBoardLogger
+from pytorch_lightning.loggers import Logger, TensorBoardLogger
 from pytorch_lightning.plugins.environments import SLURMEnvironment
 from pytorch_lightning.trainer.states import TrainerFn
 from pytorch_lightning.utilities import _TPU_AVAILABLE
@@ -909,7 +909,7 @@ def test_registries():
         pass
 
     @LOGGER_REGISTRY
-    class CustomLogger(LightningLoggerBase):
+    class CustomLogger(Logger):
         pass
 
     assert "SGD" in OPTIMIZER_REGISTRY.names
